add Collab Form and refactor AddCollab
